+++ app/domain/authentication/authn_oidc/v2/status.rb
@@ -1,0 +1,13 @@
+module Authentication
+  module AuthnOidc
+    module V2
+      class Status
+        def call(authenticator:)
+          {
+
+          }
+        end
+      end
+    end
+  end
+end