+++ app/db/repository/authenticator_repository.rb
@@ -1,10 +1,16 @@
 module DB
   module Repository
     class AuthenticatorRepository
-      def initialize(data_object:, resource_repository: ::Resource, logger: Rails.logger)
+      def initialize(
+        data_object:,
+        resource_repository: ::Resource,
+        logger: Rails.logger,
+        pkce_support_enabled: Rails.configuration.feature_flags.enabled?(:pkce_support)
+      )
         @resource_repository = resource_repository
         @data_object = data_object
         @logger = logger
+        @pkce_support_enabled = pkce_support_enabled
       end
 
       def find_all(type:, account:)
@@ -14,7 +20,14 @@ module DB
             "#{account}:webservice:conjur/#{type}/%"
           )
         ).all.map do |webservice|
-          load_authenticator(account: account, id: webservice.id.split(':').last, type: type)
+          service_id = service_id_from_resource_id(webservice.id)
+
+          # Querying for the authenticator webservice above includes the webservices
+          # for the authenticator status. The filter below removes webservices that
+          # don't match the authenticator policy.
+          next unless webservice.id.split(':').last == "conjur/#{type}/#{service_id}"
+
+          load_authenticator(account: account, service_id: service_id, type: type)
         end.compact
       end
 
@@ -27,7 +40,7 @@ module DB
         ).first
         return unless webservice
 
-        load_authenticator(account: account, id: webservice.id.split(':').last, type: type)
+        load_authenticator(account: account, service_id: service_id, type: type)
       end
 
       def exists?(type:, account:, service_id:)
@@ -36,8 +49,12 @@ module DB
 
       private
 
-      def load_authenticator(type:, account:, id:)
-        service_id = id.split('/')[2]
+      def service_id_from_resource_id(id)
+        full_id = id.split(':').last
+        full_id.split('/')[2]
+      end
+
+      def load_authenticator(type:, account:, service_id:)
         variables = @resource_repository.where(
           Sequel.like(
             :resource_id,
@@ -54,7 +71,14 @@ module DB
             args[variable.resource_id.split('/')[-1].underscore.to_sym] = variable.secret.value
           end
         end
+
         begin
+          if @pkce_support_enabled
+            allowed_args = %i[account service_id] +
+                          @data_object.const_get(:REQUIRED_VARIABLES) +
+                          @data_object.const_get(:OPTIONAL_VARIABLES)
+            args_list = args_list.select{ |key, _| allowed_args.include?(key) }
+          end
           @data_object.new(**args_list)
         rescue ArgumentError => e
           @logger.debug("DB::Repository::AuthenticatorRepository.load_authenticator - exception: #{e}")